+++ functions/ProgramTitle.fnc.php
@@ -11,7 +11,11 @@
  *
  * @example DrawHeader( ProgramTitle() );
  *
- * @global array  $_ROSARIO Sets $_ROSARIO['HeaderIcon'], uses $_ROSARIO['Menu']
+ * @since 12.0 Only load module's Menu.php file
+ *
+ * @global Set $_ROSARIO['HeaderIcon']
+ * @global $RosarioModules
+ * @global $RosarioCoreModules
  *
  * @param  string $modname  Specify program/modname (optional).
  *
@@ -19,7 +23,9 @@
  */
 function ProgramTitle( $modname = '' )
 {
-	global $_ROSARIO;
+	global $_ROSARIO,
+		$RosarioModules,
+		$RosarioCoreModules;
 
 	if ( empty( $modname ) )
 	{
@@ -33,35 +39,97 @@ function ProgramTitle( $modname = '' )
 		return ParseMLField( Config( 'TITLE' ) );
 	}
 
-	// Generate Menu if needed.
-	if ( ! isset( $_ROSARIO['Menu'] ) )
+	$module = mb_substr( $modname, 0, mb_strpos( $modname, '/' ) );
+
+	// Get right module for Teacher Programs.
+	if ( mb_substr( $modname, 0, 25 ) === 'Users/TeacherPrograms.php'
+		&& mb_strlen( $modname ) > 25 )
 	{
-		require_once 'Menu.php';
+		// Use max() to fix PHP fatal error mb_strpos(): $offset must be contained in $haystack
+		$module = mb_substr( $modname, 34, mb_strpos( $modname, '/', 34 ) - 34 );
+
+		if ( User( 'PROFILE' ) === 'teacher' )
+		{
+			$modname = mb_substr( $modname, 34 );
+		}
 	}
 
-	// Loop modules.
-	foreach ( (array) $_ROSARIO['Menu'] as $modcat => $programs )
+	if ( empty( $RosarioModules[ $module ] )
+		|| strpos( $module, '..' ) !== false )
 	{
-		// Modname not in current Module, continue.
-		if ( ! isset( $programs[ $modname ] ) )
+		// Module not found!
+		return 'RosarioSIS';
+	}
+
+	if ( isset( $RosarioCoreModules[ $module ] ) )
+	{
+		require 'modules/' . $module . '/Menu.php';
+	}
+	else // Add-on.
+	{
+		set_error_handler( function( $errno, $errstr, $errfile, $errline )
+		{
+			throw new ErrorException( $errstr, $errno, 0, $errfile, $errline );
+		} );
+
+		try
+		{
+			// Performance: up to 10% faster compared to loading root Menu.php.
+			require 'modules/' . $module . '/Menu.php';
+		}
+		catch ( ErrorException $ex )
+		{
+			/**
+			 * Old Menu.php throws a PHP fatal error
+			 * Load core modules Menu.php in case add-on adds entries to existing modules.
+			 *
+			 * @deprecated since May 2024
+			 *
+			 * @link https://stackoverflow.com/questions/8261756/how-to-catch-error-of-require-or-include-in-php
+			 */
+			foreach ( $RosarioCoreModules as $core_module )
+			{
+				require 'modules/' . $core_module . '/Menu.php';
+			}
+
+			require 'modules/' . $module . '/Menu.php';
+		}
+
+		restore_error_handler();
+	}
+
+	$profile = User( 'PROFILE' ) === 'student' ? 'parent' : User( 'PROFILE' );
+
+	// Loop programs.
+	foreach ( (array) $menu as $modcat => $menu_module )
+	{
+		if ( empty( $menu_module[ $profile ] ) )
 		{
 			continue;
 		}
 
-		// Set Header Icon.
-		if ( ! isset( $_ROSARIO['HeaderIcon'] )
-			|| $_ROSARIO['HeaderIcon'] !== false )
+		foreach ( $menu_module[ $profile ] as $program => $title )
 		{
-			// Get right icon for Teacher Programs.
-			if ( mb_substr( $modname, 0, 25 ) === 'Users/TeacherPrograms.php' )
+			if ( $program !== $modname )
 			{
-				$_ROSARIO['HeaderIcon'] = mb_substr( $modname, 34, mb_strpos( $modname, '/', 34 ) - 34 );
+				continue;
 			}
-			else
-				$_ROSARIO['HeaderIcon'] = $modcat;
-		}
 
-		return $programs[ $modname ];
+			// Set Header Icon.
+			if ( ! isset( $_ROSARIO['HeaderIcon'] )
+				|| $_ROSARIO['HeaderIcon'] !== false )
+			{
+				$_ROSARIO['HeaderIcon'] = $modcat;
+
+				// Get right icon for Teacher Programs.
+				if ( mb_substr( $modname, 0, 25 ) === 'Users/TeacherPrograms.php' )
+				{
+					$_ROSARIO['HeaderIcon'] = mb_substr( $modname, 34, mb_strpos( $modname, '/', 34 ) - 34 );
+				}
+			}
+
+			return $title;
+		}
 	}
 
 	// Program not found!